MDE/PKFE-32 changes to toolbar
