+++ app/front-end/src/features/editor/components/toolbarView/toolbarGroupButtons/index.ts
@@ -1,0 +1,3 @@
+export { ApplyGroupButtons } from './applyGroupButtons';
+export { DownloadGroupButtons } from './downloadGroupButtons';
+export { MergeGroupButtons } from './mergeGroupButtons';
