+++ app/front-end/src/features/editor/components/toolbarView/toolbarGroupItem.tsx
@@ -1,3 +1,4 @@
+import { useStatusContext } from '@/hooks';
 import { SvgIconComponent } from '@mui/icons-material';
 import { alpha, Box, Button, useTheme } from '@mui/material';
 
@@ -5,7 +6,6 @@ export interface ToolbarGroupItemProps {
   group: string;
   icon: SvgIconComponent;
   label: string;
-  disabled: boolean;
   onClick: () => void;
 }
 
@@ -35,14 +35,15 @@ export interface ToolbarGroupItemProps {
  * @param {Function} onClick - The function to be called when the button is clicked.
  * @returns {JSX.Element} The rendered Button component with an icon and label.
  */
-export const ToolbarGroupItem: React.FC<ToolbarGroupItemProps> = ({ icon: Icon, label, disabled, onClick }) => {
+export const ToolbarGroupItem: React.FC<ToolbarGroupItemProps> = ({ icon: Icon, label, onClick }) => {
   const Theme = useTheme();
+  const { blocked } = useStatusContext();
 
   return (
     <Box sx={{ height: '40%', alignContent: 'center' }}>
       <Button
         startIcon={<Icon sx={{ color: Theme.palette.text.primary }} />}
-        disabled={disabled}
+        disabled={blocked}
         onClick={() => onClick()}
         sx={{
           color: Theme.palette.text.primary,
